lead-mining-engine/requirements.txt
# Lead Mining Engine — all dependencies

# ── Web framework ────────────────────────────────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
python-multipart==0.0.20

# ── HTTP client ──────────────────────────────────────────────────────────────────
httpx==0.27.2

# ── Database ─────────────────────────────────────────────────────────────────────
asyncpg==0.30.0
chromadb==0.5.23
sentence-transformers==3.3.1

# ── Browser automation ───────────────────────────────────────────────────────────
playwright==1.49.0

# ── Reddit API ────────────────────────────────────────────────────────────────
praw==7.8.1

# ── AI enrichment ────────────────────────────────────────────────────────────────
google-generativeai==0.8.3

# ── Configuration ────────────────────────────────────────────────────────────────
pyyaml==6.0.2
python-dotenv==1.0.1

# ── Utilities ────────────────────────────────────────────────────────────────────
tenacity==9.0.0
